refactored redis client structure
