+++ packages/service-library/src/servicelib/redis/__init__.py
@@ -1,14 +1,23 @@
-from ._client import RabbitMQClient
+from ._client import RedisClientSDK
+from ._clients_manager import RedisClientsManager
+from ._decorators import exclusive
+from ._distributed_locks_utils import start_exclusive_periodic_task
 from ._errors import (
     CouldNotAcquireLockError,
     CouldNotConnectToRedisError,
     LockLostError,
 )
+from ._models import RedisManagerDBConfig
 
 __all__: tuple[str, ...] = (
     "CouldNotAcquireLockError",
     "CouldNotConnectToRedisError",
+    "exclusive",
     "LockLostError",
+    "RedisClientSDK",
+    "RedisClientsManager",
+    "RedisManagerDBConfig",
+    "start_exclusive_periodic_task",
 )
 
 # nopycln: file
